requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
urllib3
asyncio
lxml
python-slugify
scrapingant-client[async]
azure-storage-blob
azure-core
aiohttp
httpx